MDEV-18976: Implement a CHECKSUM redo log record for improved validation
New OPTION CHECKSUM redo record is added. The records are generated on
mtr_t::commit() for each modified page, CRC is counted for the page and
added in redo log record. When the record is applied, the CRC is
compared to the counted page CRC, and warning is issued in error log if
it does not match.

The new option redo_log_checksum is added, the option turns on the above
functionality.

The records are not generated for freed pages within the mtr. For this
puprose the code for pages scrubbing is reused. It collects freed page
ids within the mtr. The code was moved from buf_page_free() to
mtr_t::free() because there can be the cases when mtr_t::free() is
called without buf_page_free(), for example, when free segments are
purged in purge coordinator thread. The code is also added in
fseg_free_extent() because it invokes buf_page_free() without mtr_t::free().