MDEV-18976: Implement a CHECKSUM redo log record for improved validation
We can't use use_full_checksum parameter of
buf_flush_init_for_writing(), see previous commit message for details.

But if this parameter is no set, crc is written to 4 bytes with page_size-8
offset of a page. But encrypted pages uses full_crc32 algoritm, it
writes crc at page_size-4 offset. So if we invoke
buf_flush_init_for_writing() and the crc is written at page_size-8
offset, this leads to page checksum check error when page is read.

That is why new parameter of buf_flush_init_for_writing() is
implemented, it's the pointer to checksum to output, if this pointer is set,
buf_flush_init_for_writing() does not change page content and returns
checksum with this pointer.